Fix Understand's peculiarity with C projects
SciTools Understand expects even C projects to be created with the
language as C++. However, since the language of a project is obtained
from the project service, modifying it at the source would mean
misrepresenting the facts. For instance, Linux is a C project and
changing it to be a C++ project would be lying. Therefore, the fix is to
cast C to C++ only for the sake of appeasing Understand.