Use repository directly instead of proxy
The way RepositoryProxy is using Redis to cache objects is incorrect
because the JSON to be cached is being dumped into string and then the
string is being set as the value for a particular key. However, if the
string is larger than 512 MB which, in most cases, will be, Redis fails
to cache it. For now, the use of cache is being bypassed until a
solution is found.