Workaround an issue with incorrect git output
In a particular scenario, one of the lines of output from git log does
not conform to the expectation that the format specifier specifies. The
offending output is included below. As a workaround, we only consider
the first four fields in the output, ignoring anything that may be
after the fourth field. Clearly, the fix is merely a workaround until we
actual find a sustainable way to fix this issue permanently.

'113632d00acb569420b14eb7575833ac7e2eb311,1191964691,\"Talpey,
Thomas\,Thomas.Talpey@netapp.com'

The commit is from the Linux git repository.