Remove the validation of metrics from Understand
The Understand Python API does not seem to work as expected; the call to
get the metrics from an open Understand database should return a list of
metrics that the database has but it does not return an exhaustive list.
Since we were using this list to validate input from the user, metrics
that were valid for a particular project were rejected as invalid
because metrics() did not return it from the database. Rather than
attempting to debug the underlying issue, I am choosing to remove the
validation of the metric input.