Refactor code to use git log parser class
In addition to using a purpose-built class to parse the output from git
log, the refactoring also uses a pool of green threads from the eventlet
package to process the parsed output with a certain degree of
parallelism. The performance gain of using the pool of threads is not
marked but the option of using multiple processes to process the log
output is prevented by nameko's dependency on eventlets and green
threads.