Add custom configuration for RabbitMQ
The custom configuration is needed to increase the default Heartbeat
timeout that RabbitMQ server sets from 60 seconds to 6000 seconds. The
high timeout is necessary because some of the services could take quite
some time to finish processing the information that is needed to be then
put onto the queue. Although a workaround, I do not suspect getting rid
of this workaround anytime soon.