Add service method to get the version of a project
The version of a project is the version of its repository. The need to
retrieve version of a project becomes apparent when dealing with
caching. For instance, the understand service creates a database file
which contains the metrics that it computes from the source code in the
repository. The way the understand service is implemented is that the
database file is not recreated if one exists. As a result, if we change
the version of the repository, the database file that understand uses
may not be consistent with the repository. Using the version of the
project, we can tie the database file to the version instead of project
alone.