Include repository version in the name of UDB file
The implementation of the service cached the understand database file
from previous executions. However, if the source code of a project is
changed, the database file is neither recreated nor updated. The change
includes the version of the repository at which the database file was
created to overcome this implementation limitation.