Replace Redis with DiskCache as the cache backend
Despite using ReJSON (a variant of Redis with RedisJSON data type added
as a native type), the data we are trying to cache was apparently too
large to cache in Redis. The word "apparently" is used intentionally
because the exception that Redis raises says connection refused with no
further explanation. However, on searching for the error message, the
predominant factor in all reports of the error is an attempt to persist
a large data item to Redis. We also tried to use memcache with no
success. As a compromise, disk based cached is being introduced to at
least overcome the performance limitations that we are currently facing.