Remove redis as a service
In addition to the removal of the redis service, there were a few house
keeping tasks performed the most notable of which was the removal of
docker volumes in favor of simple directory binding. The use of
directory binding simplifies updating data in the mapped directories
that the services use. For instance, by having the repositories in a
mapped directory, changing the version of the repository no longer
requires the need to start a temporary docker container simply to run a
git checkout command.