Use yield to stream output through the layers
The refactor is in preparation for moving the implementation of the
service from nameko RPC (over RabbitMQ) to GRPC to provide service
endpoints capable of streaming their output. In the current
implementation, any attempt to return data over 2GB fails with an
exception from RabbitMQ because RabbitMQ has a hardcoded limit of 2GB on
the message size. All attempts to look for alternatives to RabbitMQ that
provides the ability to have larger messages were futile. The decision
to use GRPC to implement the services was made with both frustration
with nameko RPC over RabbitMQ and desire to implement the service in a
way that is accessible to a wider array of programming languages.