Add class to mediate running of git commands
The runner.Runner class is an intermediate point of control in the
execution of git commands. The addition of the class is the first step
in refactoring the service to cache the results of the command execution
instead of the output from the repository model. The caching of output
from the commands themselves simplifies the implementation of the rest
of the service and also improves performance of some service methods
which require two or more git commands to be executed to produce their
results.