Acquire a lock when caching data
Since the repository service is a microservice which is likely to be
deployed within a docker container, it is conceivable that the service
may be replicated. As a consequence, we have to ensure that multiple
instances of the service do not try to write the same data to the cache
at the same time. One of the prescribed recipes to prevent a collision
due to simultaneous access is to wrap the diskcache.Cache instance by a
diskcache.Lock (See http://www.grantjenks.com/docs/diskcache/api.html#
diskcache.Lock for more information).