Ignore renames when enumerating files
When using the `--name-only` flag with `git log`, the name of a file
before a move is not included in the list. For instance, if a file
`foo.c` was renamed to `bar.c`, `git log --no-merges --pretty=
--name-only` will return `bar.c` only. However, `foo.c` was a file that
was in the repository at some point and should, therefore, be included
in the listing of all files. The `--no-renames` flag to `git log` will
include both files in the listing because `git` will not attempt to
identify renames treating a rename as a file being deleted and another
file being added.