Simplify path handling from SciTools Understand
SciTools Understand behaves inconsistently when a project directory is
added to the database using its absolute path. If the root directory has
source code files, then `(root)` is prepended to all paths stored in the
database, where `(root)` is the name of the directory containing the
source code. If the root directory does not have any source code files,
then `(root)` is not prepended. For instance, if `/home/user/code/` is
added to an Understand database, all paths in the database will begin
with `code/` if `/home/user/code` has any source code files, otherwise,
`code/` is not prepended to the paths.

The inconsistency makes it difficult to match file paths returned by
Understand to that returned by git. The simplification being introduced
is this change is actually a workaround where the Understand database is
created relative to the source code location so Understand prepends all
paths with the literal `RELATIVE:/` making it easier for replacement.
However, the problem with creating the database in the source code
directory is the potential for deletion of the database when `git clean`
is run in the source code repository. So, the Understand database is
moved to a different location after it has been created and the files in
it have been analyzed.