Remove contributor metric service
The contributor metric is a measure of the number of unique contributors
to a project. The meric, being defined at the project level, is unlikely
to yield any useful insights for the developers contributing changes to
the project which may lead them to discover vulnerabilities. For
instance, informing the developer that a project has n unique
contributors is unlikely to provide any useful insights about the
change that they are contributing or the files that they are modifying
in the change.