Add service to collect the past changes metric
The past changes metric is a measure of the amount, expressed as number
of commits, a file being modified in a change has undergone in the past.
For instance, consider the output from `git log` executed against a
fictitious repository.

$ git log --numstat --reverse --pretty=%H
eaba98eac52edfe8df31148c87fb98c97e7205b6
21      0       LICENSE
1       0       README.md

17b341990d031893ccd9a3707f512d314c4d18b7
150     0       .gitignore
10      0       README.md

The past changes to file LICENSE and README.md in eaba98e would be 0
because this is the commit that introduced these files. In 17b3419,
however, the past changes of file .gitignore will be 0 because it was
introduced in this commit but the past changes of file README.md will be
1 because it was modified (in this case, added) in one past commit.