Add service to collect the past authors metric
The past authors metric is a measure of the number of developers who
have contributed a change in the past to a file being modified. Although
the authors who propose the metric do not explicitly state that the
developer of the current commit should not be included, the
implementation of the metric here does exclude the developer of the
current commit from the number of past developers. In other words, the
metric is a essentially the number of "other" developers who have
contributed a change to a particular file in the past.

As an example, consider the out from `git log` executed against a
fictitious repository.

$ git log --numstat --reverse --pretty=%H,%aN (%aE)
eaba98eac52edfe8df31148c87fb98c97e7205b6,John Smith (jsmith@email.com)
21      0       LICENSE
1       0       README.md

17b341990d031893ccd9a3707f512d314c4d18b7,Paul Smith (psmith@email.com)

150     0       .gitignore
10      0       README.md

bae7ed744a98bc8ceaf62256fb61e92d39b5be73,John Smith (jsmith@email.com)

1       0       README.md

faaac684d18bc61d791ac009d4104441e866d4eb,Eric Smith (esmith@email.com)

0       1       README.md

The past authors of file LICENSE AND README.md in eaba98e would be 0
because this is the commit that introduced these files. In 17b3419,
however, the past authors of file .gitignore will be 0 because it was
introduced in this commit but the past authors of README.md will be 1
because it was modified by one other developer in the past. Similarly,
in bae7ed7, the past authors of the file README.md will be 1 because it
was modified by one other developer. In faaac68, however, the past
authors of the file README.md will be 2 because it was modified by two
other developers in the past.