Add service to collect the message tokens metric
The message tokens metric is essentially a list of tokens in a commit
message identified by splitting the commit message at spaces or
newlines. In the paper that proposed this metric [1], the tokens were
mapped into a higher dimensional space using a vectorizer (much like the
CountVectorizer supported by scikit-learn) such that each commit was
represented by a high-dimensional vector in which the value of a
dimension was 1 if the commit had the token associated with the
dimension, 0 otherwise.

In the implementation of the metric, rather than returning the higher
dimensional vector representation of each commit message, we return a
sparse representation. In this sparse representation, a list of all
tokens across all commit messages is returned as part of the response.
The tokens are essentially the dimensions of the vector space to which
the tokens in the commit message will be mapped to. With the dimensions
established, the tokens in a commit message are returned as indices
(list of integers) that map to the dimensions such to indicate that the
dimension indexed by the token index is in the commit message.

References

[1] VCCFinder: Finding Potential Vulnerabilities in Open-Source Projects
    to Assist Code Audits