Add method to get content of a git object
A git object is essentially an internal representation of a file in a
git object database. Each git object is uniquely identified by a SHA-1
checksum hash that is assigned to the object when the object (i.e. a
file) is persisted (i.e. committed) to the git object database (i.e.
repository). Each commit, therefore, is a collection of pairs of git
object identifiers, with each pair having one object identifier to
identify the version of a file before the commit is applied and one
object identifier to identify the version of a file after the commit is
applied.

The `git diff-tree -r <sha>` command shows the identifiers of the
object(s) affected (added, deleted, or modified) by a commit. For
example, consider the excerpt of the output of the `git diff-tree -r
<sha>` for a particular commit in the psf/requests project.

$ git diff-tree -r 4bad52caa224a83b3e49b2063fcb21e7b85bbae9
4bad52caa224a83b3e49b2063fcb21e7b85bbae9
...
:100644 000000 e4d148b3c5b995f810428ce6e44643147d406b15 0000000000000000000000000000000000000000 D      requests/packages/chardet/constants.py
:100644 100644 ff4272f82a0bad7585a6130f089618d09f34f02c efd793abca4bf496001a4e46a67557e5a6f16bba M      requests/packages/chardet/cp949prober.py
:000000 100644 0000000000000000000000000000000000000000 04512072251c429e63ed110cdbafaf4b3cca3412 A      requests/packages/chardet/enums.py
...

The fourth and fifth columns in the output shown above show the git
object identifiers of pre- and post-commit versions of corresponding
files. Naturally, a file that is added (like enums.py above) will have
all zeros in the pre-commit object identifier (indicating an empty file)
and a file that is deleted (like constants.py above) will have all zeros
in the post-commit object identifier.

The method to get content of a git object being introduced in this
change uses `git show` to retrieve the content of a the git object
identified by an object identifier.