Add method to get deltas of a repository
The addition of the get_deltas method is in preparation to modify the
get_changes method to strictly return specific information about the
change such as the type of change (addition, deletion, modification,
etc.) and identifiers of the object in the repository before and after
the change. In this strict view, the number of lines inserted and
deleted is not relevant to the change itself.