Use git diff-tree to get repository changes
The use of git diff-tree to get the changes to a repository instead of
using git log is made necessary by the need capture the identifiers of
version of a file changed in a commit. The identifier (called oid in
git) uniquely identifies an object is the repository and can be used to
retrieve a particular version of a particular file.

The need for object identifiers was made necessary by the need to
collect the function churn metric which requires two version of a source
code file (before and after a commit) to be parsed to compare the
functions that were added, deleted, and modified.