Add service to parse source code
The `parser` service will wrap around multiple parsers (CParser
implemented with clang as of this commit) to parse source code. The
utility of this service, at the moment, is to identify functions that
are defined in the source code to aid in determining the changes to
functions between commits to a subject source code repository.