Use project service to get project object
The change is in preparation to integrate the churn service with the
newly introduced parser service to collect the function churn metric.
The parser service is only implemented for a subset of the languages
(C/C++ as of this writing) and having the language of the project will
enable us to check if the parser can parser the project before
attempting to invoke the service.