Move configuration of libclang path to environment
The use of set_library_path was raising an exception when the parser
service was run in a replicated mode. The change removes the hardcoding
of the path to the location of the libclang library from the source
code. The configuration of the library path is now done using the
LD_LIBRARY_PATH environment variable which can be done fairly easily in
the docker-compose file (as is done here).