Introduce function insertions and deletions
The change is a part of the implementation of the function churn metric.
The change introduces the ability to collect two measures of the
function churn metric: number of functions inserted and number of
functions deleted in a commit. The remaining measure of function churn
is number of functions *modified* in a commit the implementation of
which will be in a separate change to reduce churn of introducing the
metric.