Use IntEnum instead of Enum to improve usability
Instances of types inheriting from IntEnum can be directly compared to
integers. For instance, a comparison of the form `1 == ChangeType.ADDED`
is not permitted if ChangeType inherited from Enum. The need for
comparison with integers is due to the type conversion of the
enumeration to an integer when transmitting the response to the service
consumer. The IntEnum provides consumers the ability to compare the raw
integer value rather than having to convert the integer to the
enumeration type before comparing.