Fix bug in the way functions were identified
When a source code file is parsed using libclang, the output is
essentially an astract syntax tree with the root node representing the
entire source code file. The initial implementation of the get_functions
method *did not* recursively parse the entire abstract syntax tree but
only the children of the root node (i.e. only those nodes that were at
the second level of the abstract syntax tree). If all functions in a
source code file were directly defined in the file itself (i.e. not
within a namespace), the bug would not have been caught. However, in
testing, there were source code files in which the functions were indeed
defined inside of a namespace and these functions were not returned by
get_functions. The change implements a breadth first traversal of the
entire abstract syntax tree to fix the bug.