Add method to get lines changed in a commit
The lines changed in a commit are returned as a list of line numbers
that were inserted and a list of line numbers that were deleted in a
commit. Since there is not direct way of getting the line number of
lines modified in a commit using the `git` command line. While the
hunk headers in the output from `git show --patch <commit>` could be
used to *compute* the number of the lines that were modified in a
commit, the likelihood of inadvertent mistakes in computing the line
number is not insignificant. As a convenient alternative, pygit is
introduced and used to get the lines changed in a commit.