Add service method to query if a file is parsable
The service method is essentially a convenience feature being
implemented to avoid the need to get contents of a file from the
repository if the file is not parsable (i.e. no parser exists for the
file). The service consumer is responsible to use the service method to
check if a path is parsable before invoking a method to actually parse
the file.