Add modifications to the function churn metric
The modifications in the function churn metric is the number of
functions modified in a commit. The approach to collect the metric
extends the approach used to collect the number of functions inserted
and number of functions deleted metrics by determining the overlap in
the lines modified in a commit with the lines that definitions of
functions span.