Remove function churn from churn
The decision to include function churn as a variant of churn proved to
complicate the implementation of the service. Furthermore, the inclusion
of function churn with churn (rather, line churn) meant that we lost the
ability to collect line churn without collecting function churn and, as
of this change, the implementation of function churn is quite not as
performant as one would like. The removal of function churn means that
the churn metric can be collected independently aiding improved
flexibility.