Remove unnecessary argument `processes`
The `processes` argument to the service methods violates the separation
of concerns because the service consumer should not be able to control
the number of processes that a service spawns to fulfil a request. The
number of processes that the service uses should be an internal
attribute of the service the configuration of which should further be
delegated to the deployment environment (docker in this case). The
service should be greedy and use as many processes as there are
processors in the deployment environment.