Handle None from repository.get_content
The get_content method of the repository service can return None in
cases when an object identifier is not in the repository. The only known
case of this happening is when attempting to get the before version of a
added file or when attempting to get the after version of a deleted
file.