Use Redis to cache content of git objects
The introduction of Redis to cache git objects is a desperate attempt at
mitigating two concerns: (1) Dependency on disk when getting content of
git object which may block (2) Apparent memory issue with pygit which
seems to not clean up after itself. By caching all objects in a
repository in Redis, we mitigate both these concerns at the expense of
additional memory.