Use Redis to cahce components of function churn
The introduction of Redis to cache the components (insertions,
deletions, and modifications) of the function churn metric is an attempt
to improve the performance of the function churn metric service. Despite
using Redis to cache the git objects that are used to compute the
function churn components, the performance was not as expected. However,
with the use of the cache the performance seems to have improved
considerably.