Fix bug with RELATIVE:/ in name of file
The change to create the Understand database relative to the location of
the source code from which the metrics were being collected was made
necessary by paths being absolute. However, in impelementing the change,
the replacement of the prefix---RELATIVE:/---that Understand adds was
missed for name of a file causing lookup failures downstream.