Use environment variables to ease maintenance
The use of environment variables, in addition to aiding maintenance,
also avoids having sensitive information in the docker compose file. In
our workflow, we use `direnv` to load environment variables declared in
an .envrc file and `envsubst` to replace environment variables in the
docker-compose.yml file with values from the environment before using it
to deploy services. The actual command used in our workflow is
`envsubst < docker-compose.yml | docker stack deploy -c - <service>`.