Remove --no-merges from git commands
As a consequence of the `--no-merges` switch to git commands, few files
that were indeed present in the repository were not being reported as
being part of the repository leading to issue when attempting to load
metrics associated with these files. The root cause of the missing files
was traced back to the files having been added as part of a merge commit
and since the command used to get "all" files ever tracked by the
repository used the `--no-merges` switch, these files were not being
reported.