Add similarity index as a parameter to get_moves
The similarity index is a metric that git computes to determine if a
file being added in a change is a rename of a file being removed in the
same change. For instance, if foo.c is renamed to bar.c in a change, git
tracks it as foo.c being removed and bar.c being added. However, when
presenting the history, git attempts to detect renames by computing the
similarity of the contents of foo.c and bar.c and uses a threshold to
consider a file addition/removal pair as a rename. The `similarity`
parameter to `get_moves` allows service consumers to specify the
threshold for the similarity index.