Fix bug with merge commits not being returned
By default, the `git diff-tree` command does not show changes for a
merge commit. The bug was with the command and the missing `-m` option
which makes `git diff-tree` show changes for a merge commit. The problem
with `-m` is that for merge commits with more than one parent, the
output is shown for each pair of commit and its parent. The command
provides no way of specifying that only the first parent (i.e. parent
from the branch into which the commit is being merged) should be
considered. The use of `awk` is to overcome this limitation by
explicitly passing the commit and parent hash pairs to `git diff-tree`.