Fix bug in function churn for added/deleted files
TL;DR Function churn was not collected from changes in which a file was
added or deleted.

An oversight in the implementation of a condition used to determine the
existence of functions parsed from two versions of a file (before and
after a commit) led to function churn not being collected from changes
where a file was added or deleted. The condition prevented the
collection of the function churn metric when one of the list of
functions before and after the commit was None. If a file is added, then
the before version of the file will not have functions. Similarly, if a
file is deleted, then the after version of the file will not have
functions.

The change fixes the bug to attempt to collect the function churn metric
when at least one of the lists of functions (before and after) is not
None. The only scenario where both lists of functions are empty is when
the file is not parseable (i.e. there is no parser supported for the
language in which the file is written in) or the parsing of the file
failed.