Fix performance issue caused by uncached data
The caching of raw function churn components (i.e. insertions,
deletions, and modifications) was nested within the if conditional if at
least one version of the file (before or after a commit) was
successfully parsed. The performance implication of not caching the
components of files that were not parsed is that the next time around
the service would attempt to parse these files again. The change fixes
this issue by caching everything. As a direct consequence of the change,
the default value for the components can no longer be (None, None, None)
because Redis does not understand None. We now use (-1, -1, -1) as the
default value because -1 is an invalid value for churn anyway. We do,
however, translate -1 back to None so the service consumer need not be
made aware of the integration with Redis.