Set license code before running service
The licensing service that Scitools uses seems to have be allowing the
understand service to use the API without having to set a license.
However, it appears that the apparent flaw in the licensing service
seems to have been resolved which caused the understand service to fail
with the NoApiLicense exception. The change sets up the license (stored
in an environment variable) before starting the service.