Rename rabbitmq.conf to rabbitmq.template.conf
The rename is in preparation to set the default user and password
of a RabbitMQ in the configuration file. Although the RabbitMQ image
supports RABBITMQ_DEFAULT_USER and RABBITMQ_DEFAULT_PASS environment
variables, we cannot seem to be able to use both a configuration file
and these environment variables at the same time. The `entrypoint.sh`
script associated with the image apparently modifies the configuration
file at runtime to incorporate the environment variable values into the
configuration file. However, when a configuration file is also provided,
the `entrypoint.sh` script fails to modify the configuration file at
runtime due to a permissions issue. The solution is to use the
configuration file to configure the default user and password and since
this is sensitive information, a template is being created with the
actual configuration file configured to be ignored by git.