Use images from GitHub Package Registry
During initial development, the images pushed to a private docker
registry hosted locally. However, in preparation for the release of the
services, the images have all been pushed to GitHub Package Registry.
The change updates the docker-compose.yml file to use the images from
the GitHub Package Registry instead of the local private registry.