Use ~= to allow compatible release in requirements
The use of >= to specify expected version of dependencies is incorrect
because >= would match even when the major version of the dependency is
changed. In semantic versioning scheme [1], a major version change
indicates breaking changes to the API and by allowing dependencies with
major version change could break the services. Fortunately, the
comatible release operator (~=), specified in PEP 440 [2], allows a
dependency to be updated with any minor or patch changes but prevents
any major changes.

References

[1] https://semver.org
[2] https://www.python.org/dev/peps/pep-0440/#compatible-release