Fix bug in the specification of diskcache version
The diskcache version was specified using >= instead of ~= before a
recent change to replace >= with ~=. As a consequence, the version of
diskcache that was installed was 4.1.1 and not 3.1.x. The container
images created after the change from >= to ~= failed to run because Lock
was not defined in the diskcache module.

The change resolves the bug by specifying the version of diskcache that
is actually used by the repository service in the most recently
published image.

The bug is a cautionary tale against the use of >= when specifying
versions in requirements.txt.