Add method to get commits that last modified lines
The commit that last modified a line of a file is the commit that is
blamed by `git blame` relative to a commit provided as input. The
assumption is that the lines for which the last modifying commit is to
be determined were changed in the commit provided as input.