Fix bug in parsing `git blame` output
The fields in each line of output from `git blame` upto the, but
excluding the actual line of code, was _assumed_ to be delimited by
spaces. However, this assumption was shown to be invalid when attempting
to identify commits that last modified line 76 of a file in Apache
Tomcat, java/org/apache/catalina/core/LocalStrings.properties, relative
to bc0f529a1c7a7acb568588371ff890511f5baf9d. For some inexplicable
reason, the email address associated with the commit that last modified
line 76 has a space in it.

The change fixes the bug by parsing the line in search of the `)`
character that indicates the end of `git blame` information and the
beginning of the line of code being blamed. The change is stopgap until
the issue described in #10 is resolved.