Use python:3.7-slim-buster as base to reduce size
The change switches the base image of all services in the samaritan
portfolio to `python:3.7-slim-buster` to considerably lower the size of
the final image.

In addition to the change in base image, the Dockefiles have been
refactored to incorporate best practices enforced by Haskell Dockerfile
Linter (See https://github.com/hadolint/hadolint).