Escape names and extensions in `languages.yml`
`languages.yml` maps a `names` and `extensions` to a programming
language. However, if the name or extension has a character that has a
special meaning in regular expressions, the regular expression
constructed from the `names` and `extensions` will not match as
intended. For instance, if `c++` is an extension that maps to the C++
language, the `+` in `c++` must be appropriately escaped. The change
introduces appropriate escapes.