Fix bug with missing pos:* from block_content
The `<block_content>` element in `<block>` in `<function>` (representing
a function definition) was assumed to always have the `pos:start` and
`pos:end` attributes. However, in case of functions defined with an
empty body, the `<block_content>` element does not have `pos:*`
attributes. In such cases, the `parser` service raised a KeyError. The
change resolves the bug by checking the `<block_content>` has
attributes.

Change resolves #25.