Check <name> is in <function> or <function_decl>
While all function declarations and definitions _can_, in theory,
assumed to have a name, there are certain cases where SrcML fails to
identify the name of a function declared or defined in source code. The
change adds a check to account for the case when SrcML cannot determine
the name of a function.

Change resolves #26.