Fix bug where function names are returned as `None`
In cases where a function name has an operator in it (like when the
scope resolution operator is used in the function name), the `<name>`
element of the `<function>` or `<function_decl>` is does not contain the
textual name of the function but instead child elements that inturn form
the name of the function. The change iterates through the children of
the `<name>` element and concatenates the text of each of those elements
to form the function name.

Change resolves #27.