Remove ability to not specify certain arguments
As described in Issue #8, the metric services' `collect` methods are
being modified to allow incremental collection of metrics. As a first
step toward this change, the `collect` methods were modified by adding
optional arguments to specify `sha` of commit, `path` or file, and
`name` and `email` of developer. The arguments were optional to maintain
backwards compatibility. The change makes these optional arguments
required as we move towards accomplishing the incremental collection.