Simplify measures returned by metric services
With Issue #8, metric services are now capable of collecting metrics
from a change, commit, file, or developer at a specific point in time.
As a result, wrapping the metric measures around in a model with
additional information (such as the change for change-level metric or
the commit for commit-level metric) seemed rather unnecessary, at best,
and cumbersome to use, at worst. The change simplifies the return from
all metric services that now support incremental collection. For
instance, the ownership metric is no longer wrapped around by the
`ownership.models.Ownership` model but is instead a floating point
number.