Fix issue in getting changes from merge commits
The changes contributed to in a merge commit must be handled differently
from changes contributed to in a regular commit. The indistinction in
getting changes for regular vs. merge commits lead to Issue #36. The fix
involves considering only those changes in a merge commit that are
directly attributable to the merge commit itself. See
https://github.com/samaritan/services/issues/36#issuecomment-714690096
for a detailed description of the decision behind the fix.

Change resolves #36.