Simplify structure of data returned by `repository`
The structure of the data returned by certain service methods in the
`repository` service include metadata which may not be needed by the
service consumer. The inclusion of such metadata only increases the size
of the data on the wire. The change simplifies the structure to only
return that data requested. The change is breaking and has implications
on all services that use the `repository` service.

Change resolves #38.