Add column to comments from `parser.get_comments`
The change introduces two additional model classes: `Position` and
`Span`. The `parser.get_functions` service method is also modified to
use the `Position` and `Span` classes to communicate the span of a
function definition or declaration with the `Position.column` field
being set to None since it is not used downstream at this point.

Change resolves #41.