Make RabbitMQ protocol configurable
When running RabbitMQ as part of the Samaritan stack of services, SSL is
not enabled. However, when using RabbitMQ on Amazon MQ, the only
endpoint that is accessible is SSL enabled. As a result, the protocol
used to communicate with RabbitMQ must be configurable until the time
that we have SSL-enabled RabbitMQ be the default.