Fix KeyError issue in `get_changes`
The cause of the issue was the missed consideration for cases in which
the file has a space in its name. The change resolves the issue by
altering the parsing of output from `git diff-tree` to be left-to-right
following the format documented [1] rather than a mixture of both
left-to-right and right-to-left.

Change resolves #56.

References

[1] https://git-scm.com/docs/git-diff-tree#_raw_output_format