Handle paths that have space in their names
In the initial development of the `repository` service, there were no
files with a space in their name. As a result, the git commands did not
quote the path used. However, the discovery of Issue #56 revealed the
deeper limitation in the service when dealing with files that have a
space in their names. The change mitigates the limitation by quoting the
path.

Change resolves #57.