Return line ranges in `get_linechanges`
The `get_linechanges` method returned the lines changed in a commit. The
lines changed was returned as a list of integers which could lead to
performance problems when a file with several thousand lines of code is
deleted. In this change, the list of lines changed is collapsed to a
list of tuples representing range of lines changed.