Compress response from `get_lastmodifiers`
The initial implementation of `get_lastmodifiers` had a flaw which when
encountered could lead to the saturation of RabbitMQ memory. The flaw
was in the design: response from `get_lastmodifier` was a mapping
between a line number a commit that last modified the line. If a large
file was renamed, attempting to get the last modifiers of all lines in
the file could result in a response that may have a large number of
repeated commit objects if the same commit modified several lines.

In this change, the flaw is resolved by collecting all lines that were
last modified by a particular commit and returning a mapping between the
commit and a collapsed list of line ranges that the commit modified.